+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Remove UiProgressCircular from the DOM when `loading` is false. Previously, it was hidden with `display: none`.
 - Remove pointer cursor, to match default HTML buttons.
 - For secondary buttons, use button color as background for hover and focus.
+- Add `href` prop. Setting this prop will render an anchor tag.
 
 ## UiIconButton
 
@@ -21,6 +22,7 @@
 - Remove pointer cursor, to match default HTML buttons.
 - For secondary icon buttons, use button color as background for hover and focus.
 - Add new size: `mini`
+- Add `href` prop. Setting this prop will render an anchor tag.
 
 ### UiTabs and UiTab